// src/infrastructure/interfaces/auth.responses.ts
// Generated by https://quicktype.io

export interface AuthResponse {
  id: string;
  email: string;
  fullName: string;
  isActive: boolean;
  roles: string[];
  token: string;
}

export interface SRRAuthResponse {
  access_token: string;
  token_type: string;
  user: User;
}

export interface User {
  email: string;
  is_active: boolean;
  is_superuser: boolean;
  full_name: string;
  id: number;
}